[⚙️ chore] 클라이언트 axiosinstance와 interceptor 파일 병합 #57
- 인터셉터 파일 삭제
